dashboard/backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
flask-cors==6.0.1
Werkzeug==3.1.3

# AWS and Cloud
boto3==1.39.14
botocore==1.39.14

# HTTP and API clients
requests==2.31.0
urllib3==2.3.0

# Data processing (essential only)
pandas==2.2.3
numpy==2.2.3

# Time and date handling
python-dateutil==2.9.0.post0
pytz==2025.1

# Security and encryption
cryptography==42.0.8

# Configuration and environment
python-dotenv==1.0.0

# JSON Web Tokens for auth
PyJWT==2.10.1

# Excel/CSV processing
openpyxl==3.1.5
xlrd==2.0.2

# Amazon SP API (if needed)
python-amazon-sp-api==1.8.22

# Security
cryptography>=3.4.8

# Supporting libraries
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
six==1.17.0